Adds a randomly walking people
At this stage the people contained in a table called `people` which
is a collection of `Person` entities. Each one is initialised in a
random location with a random walking trajectory.

Position and velocity are both implemented using the vector class from
the `hump` repository [https://github.com/vrld/hump].

Next step is to make some rudimentary collision detection.