Adds shout action to the Person
Person:shout()
Cast a circular segment out in the direction the person is facing
creating a temporary field,

At the moment shout casts a semi-circular segment out from the person and
detects all the people within that segment, applying the recovered state to
them for illustration purposes.

This is accomplished using an algorithm from this answer on stackoverflow:
https://stackoverflow.com/a/13675772.

Detection algorithm
1. check if angle of point is less than angle of sector sectorStart
2. check if angle of point is greater than angle of sectorEnd
3. check if distance of point is less than sector radius
if all three checks are true the point is within the sector

TODO: any person within the field gets pushed away from the segment.
      Also draw some random ASCII characters from the set *&%$#!*