Major Revision
Moves the coordinate system to a sideways facing 2D system where y is up
and x is horizontal. I'm taking the code I wrote for the anisotropy
player and merging it with this project. The Character class is
superseding the person class.

Adds a room class, which is effectively a wrapper around a wf.world with
obstacles in it.

-- Rooms are themselves state machines!!!